backend/requirements.txt

Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.20
passlib==1.7.4
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
email-validator==2.1.0.post1
stripe==7.5.0
